+++ storage/docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Fix `list_assets` response parsing and support pagination size control with
+  the `page_limit` parameter.
+
 ## 1.0.1 - 2023-10-13
 
 - Fix support for Python 3.8